+++ tcp-gameserver/src/main/java/org/xxdc/oss/example/transport/SecureDuplexMessageHandler.java
@@ -2,15 +2,10 @@ package org.xxdc.oss.example.transport;
 
 import java.io.IOException;
 import java.lang.System.Logger;
-import java.lang.System.Logger.Level;
-import java.security.AlgorithmParameters;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.InvalidKeyException;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
-import java.security.PublicKey;
 import java.security.SecureRandom;
 import java.security.Security;
 import java.security.spec.InvalidParameterSpecException;
@@ -18,16 +13,14 @@ import javax.crypto.BadPaddingException;
 import javax.crypto.Cipher;
 import javax.crypto.DecapsulateException;
 import javax.crypto.IllegalBlockSizeException;
-import javax.crypto.KEM;
 import javax.crypto.NoSuchPaddingException;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.IvParameterSpec;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.bouncycastle.pqc.jcajce.provider.BouncyCastlePQCProvider;
-import org.bouncycastle.pqc.jcajce.spec.KyberParameterSpec;
 import org.xxdc.oss.example.security.KyberKEMProvider;
 
-public abstract sealed class SecureDuplexMessageHandler implements MessageHandler {
+public abstract class SecureDuplexMessageHandler implements MessageHandler {
 
   private static final Logger log = System.getLogger(SecureDuplexMessageHandler.class.getName());
 
@@ -39,13 +32,6 @@ public abstract sealed class SecureDuplexMessageHandler implements MessageHandle
 
   public SecureDuplexMessageHandler(DuplexMessageHandler handler) {
     this.handler = handler;
-    registerSecurityProviders();
-  }
-
-  private void registerSecurityProviders() {
-    Security.addProvider(new BouncyCastleProvider());
-    Security.addProvider(new BouncyCastlePQCProvider());
-    Security.addProvider(new KyberKEMProvider());
   }
 
   /**
@@ -172,204 +158,6 @@ public abstract sealed class SecureDuplexMessageHandler implements MessageHandle
   private void checkInitialized() {
     if (!initialized) {
       throw new IllegalStateException("SecureMessageHandler has not been initialized.");
-    }
-  }
-
-  /**
-   * Represents a secure message handler for the server side of a secure communication channel. This
-   * class extends the `SecureMessageHandler` class and is responsible for initializing the secure
-   * channel, exchanging the shared secret key with the client, and providing methods for sending
-   * and receiving encrypted messages.
-   */
-  public static final class Server extends SecureDuplexMessageHandler {
-    /**
-     * Constructs a new `SecureServerMessageHandler` instance with the given `RemoteMessageHandler`.
-     *
-     * @param remoteMessageHandler the `RemoteMessageHandler` to use for sending and receiving
-     *     messages
-     */
-    public Server(DuplexMessageHandler remoteMessageHandler) {
-      super(remoteMessageHandler);
-    }
-
-    /**
-     * Initializes the secure message handler by setting up the secure channel and exchanging the
-     * shared secret key with the client.
-     *
-     * @throws IOException if there is an error during the initialization process
-     */
-    @Override
-    public void init() throws IOException {
-      try {
-        handler.init();
-        log.log(
-            Level.DEBUG,
-            "Initializing secure channel for {0}. Exchanging shared key...",
-            getClass().getSimpleName());
-        sharedKey = exchangeSharedKey();
-        initialized = true;
-        log.log(
-            Level.DEBUG,
-            "Secure connection for {0} established with {1} shared key.",
-            getClass().getSimpleName(),
-            sharedKey.getAlgorithm());
-      } catch (NoSuchAlgorithmException
-          | NoSuchProviderException
-          | InvalidParameterSpecException
-          | InvalidKeyException
-          | InvalidAlgorithmParameterException
-          | DecapsulateException e) {
-        throw new IllegalArgumentException(
-            "Invalid security configuration/exchange: " + e.getMessage(), e);
-      }
-    }
-
-    /**
-     * Exchanges the shared secret key with the client using the Kyber key encapsulation mechanism
-     * (KEM).
-     *
-     * @return the shared secret key
-     * @throws NoSuchAlgorithmException if the specified algorithm is not available
-     * @throws IOException if there is an error during the key exchange process
-     * @throws NoSuchProviderException if the specified provider is not available
-     * @throws InvalidParameterSpecException if the parameter specification is invalid
-     * @throws InvalidAlgorithmParameterException if the algorithm parameters are invalid
-     * @throws InvalidKeyException if the key is invalid
-     * @throws DecapsulateException if there is an error during the decapsulation process
-     */
-    protected SecretKey exchangeSharedKey()
-        throws NoSuchAlgorithmException,
-            IOException,
-            NoSuchProviderException,
-            InvalidParameterSpecException,
-            InvalidAlgorithmParameterException,
-            InvalidKeyException,
-            DecapsulateException {
-      var keyPair = generateKeyPair();
-      publishKey(keyPair.getPublic());
-      // Receiver side
-      var encapsulated = handler.receiveBytes();
-      var encapsulatedParams = handler.receiveBytes();
-      var kem = KEM.getInstance("Kyber", "BCPQC.KEM");
-      var params = AlgorithmParameters.getInstance("Kyber");
-      params.init(encapsulatedParams);
-      var paramSpec = params.getParameterSpec(KyberParameterSpec.class);
-      var decapsulator = kem.newDecapsulator(keyPair.getPrivate(), paramSpec);
-      return decapsulator.decapsulate(encapsulated);
-    }
-
-    /**
-     * Generates a new Kyber key pair and publishes the public key to the client.
-     *
-     * @return the generated key pair
-     * @throws NoSuchAlgorithmException if the specified algorithm is not available
-     * @throws IOException if there is an error during the key publication process
-     */
-    private KeyPair generateKeyPair() throws NoSuchAlgorithmException, IOException {
-      var keyPairGen = KeyPairGenerator.getInstance("Kyber");
-      var keyPair = keyPairGen.generateKeyPair();
-      return keyPair;
-    }
-
-    /**
-     * Publishes the given public key to the client.
-     *
-     * @param pk the public key to publish
-     * @throws IOException if there is an error during the publication process
-     */
-    public void publishKey(PublicKey pk) throws IOException {
-      handler.sendObject(pk);
-    }
-  }
-
-  /**
-   * Represents a secure message handler for the client side of a secure communication channel. This
-   * class extends the `SecureMessageHandler` class and is responsible for initializing the secure
-   * channel, exchanging the shared secret key with the server, and providing methods for sending
-   * and receiving encrypted messages.
-   */
-  public static final class Client extends SecureDuplexMessageHandler {
-
-    /**
-     * Constructs a new `SecureClientMessageHandler` instance with the given `RemoteMessageHandler`.
-     *
-     * @param handler the `RemoteMessageHandler` to use for the secure communication channel
-     */
-    public Client(DuplexMessageHandler handler) {
-      super(handler);
-    }
-
-    /**
-     * Initializes the secure message handler by setting up the secure channel and exchanging the
-     * shared secret key with the client.
-     *
-     * @throws IOException if there is an error during the initialization process
-     */
-    @Override
-    public void init() throws IOException {
-      // Sender side
-      try {
-        handler.init();
-        log.log(
-            Level.DEBUG,
-            "Initializing secure channel for {0}. Exchanging shared key...",
-            getClass().getSimpleName());
-        sharedKey = exchangeSharedKey();
-        initialized = true;
-        log.log(
-            Level.DEBUG,
-            "Secure connection for {0} established with {1} shared key.",
-            getClass().getSimpleName(),
-            sharedKey.getAlgorithm());
-      } catch (ClassNotFoundException
-          | IOException
-          | NoSuchAlgorithmException
-          | NoSuchProviderException
-          | InvalidKeyException
-          | InvalidAlgorithmParameterException e) {
-        throw new IllegalArgumentException(
-            "Invalid security configuration/exchange: " + e.getMessage(), e);
-      }
-    }
-
-    /**
-     * Initializes the secure message handler by exchanging a shared key with the remote party. This
-     * method is called on the sender side to set up the secure communication channel.
-     *
-     * @throws IOException if there is an error initializing the communication handler
-     * @throws ClassNotFoundException if the remote public key class cannot be found
-     * @throws NoSuchAlgorithmException if the specified key exchange algorithm is not available
-     * @throws NoSuchProviderException if the specified cryptographic provider is not available
-     * @throws InvalidKeyException if the remote public key is invalid
-     * @throws InvalidAlgorithmParameterException if the key exchange parameters are invalid
-     */
-    protected SecretKey exchangeSharedKey()
-        throws NoSuchAlgorithmException,
-            NoSuchProviderException,
-            ClassNotFoundException,
-            IOException,
-            InvalidAlgorithmParameterException,
-            InvalidKeyException {
-      var kem = KEM.getInstance("Kyber", "BCPQC.KEM");
-      var publicKey = retrieveKey();
-      var paramSpec = KyberParameterSpec.kyber1024;
-      var encapsulator = kem.newEncapsulator(publicKey, paramSpec, null);
-      var encapsulated = encapsulator.encapsulate();
-      handler.sendBytes(encapsulated.encapsulation());
-      handler.sendBytes(encapsulated.params());
-      return encapsulated.key();
-    }
-
-    /**
-     * Retrieves the public key from the server
-     *
-     * @return the public key from the server
-     * @throws ClassNotFoundException if the class is not found whilst deserializing from the
-     *     message handler
-     * @throws IOException if there is an error with the communications channel
-     */
-    private PublicKey retrieveKey() throws ClassNotFoundException, IOException {
-      return (PublicKey) handler.receiveObject();
     }
   }
 }
